+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### Other Changes
 
+- Added a PythonOptions that allows the user to use an env var to declare a python venv
+
 ## 0.5.7 (2025-11-05)
 
 ### Features Added